Revert "progress 09/24"
