+++ client/LoginRegister.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import UserForm from './UserForm';
+
+const LoginRegister = () => {
+  return (
+    <div>
+      {}
+      <UserForm />
+    </div>
+  );
+};
+
+export default LoginRegister;
